% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armLevel.R
\name{annotate_integer_copy_number}
\alias{annotate_integer_copy_number}
\title{don't know}
\usage{
annotate_integer_copy_number(gene_level, AMP_thresh_tcn = 6)
}
\arguments{
\item{gene_level}{numeric first argument}

\item{AMP_thresh_tcn}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
