% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno.R
\name{annotate_maf_with_facets_cf_tcn_lcn}
\alias{annotate_maf_with_facets_cf_tcn_lcn}
\title{don't know}
\usage{
annotate_maf_with_facets_cf_tcn_lcn(maf, out, fit,
  iTumor_Sample_Barcode = NULL)
}
\arguments{
\item{maf}{numeric first argument}

\item{fit}{numeric first argument}

\item{iTumor_Sample_Barcode}{numeric first argument}

\item{diplogr}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
