% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armLevel.R
\name{armLevel_main_function}
\alias{armLevel_main_function}
\title{don't know}
\usage{
armLevel_main_function(filename, outfile, method)
}
\arguments{
\item{filename}{character vector of filenames to be processed}

\item{outfile}{character Output filename}

\item{method}{character Method used to calculate integer copy number. Allowed values cncf or em}
}
\value{
don't know
}
\description{
temporary function
}
