% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armLevel.R
\name{get_gene_level_calls}
\alias{get_gene_level_calls}
\title{don't know}
\usage{
get_gene_level_calls(cncf_files, method, WGD_threshold = 0.5,
  amp_threshold = 5, mean_chrom_threshold = 0,
  fun.rename = function(filename) {     filename })
}
\arguments{
\item{cncf_files}{numeric first argument}

\item{method}{numeric first argument}

\item{WGD_threshold}{numeric first argument}

\item{amp_threshold}{numeric first argument}

\item{mean_chrom_threshold}{numeric first argument}

\item{fun.rename}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
