% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno.R
\name{integer_cn_table}
\alias{integer_cn_table}
\title{don't know}
\usage{
integer_cn_table(out, fit, em = FALSE)
}
\arguments{
\item{fit}{numeric first argument}

\item{diplogr}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
