% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno.R
\name{main}
\alias{main}
\title{don't know}
\usage{
main(maf, cncf_files, purity_fit_filename = "")
}
\arguments{
\item{maf}{numeric first argument}

\item{facets_file}{numeric first argument}

\item{file_type}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
