% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots.R
\name{plotSampleCNCF.custom}
\alias{plotSampleCNCF.custom}
\title{don't know}
\usage{
plotSampleCNCF.custom(jointseg, out, fit, main = "")
}
\arguments{
\item{jointseg}{numeric first argument}

\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{main}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
