% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{reSeg}
\alias{reSeg}
\title{don't know}
\usage{
reSeg(orig, cval = 35)
}
\arguments{
\item{orig}{numeric first argument}

\item{cval}{numeric first argument}
}
\value{
don't know
}
\description{
To resegment previously segmented samples (by Nick), provided the Rdata file
exists of the original output. Testing the increase of cval (default=35 for
exome-wide data) for WGS data.
}
