% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armLevel.R, R/check.R, R/mergeTN.R
\name{annotate_integer_copy_number}
\alias{annotate_integer_copy_number}
\alias{af_grid}
\alias{mergeTN_function}
\title{don't know}
\usage{
annotate_integer_copy_number(gene_level, AMP_thresh_tcn = 6)

af_grid(maf, output_pdf, log = "-", threshold = 4, type = "ccf",
  purity = NA, nbins = 30, label_WGD = "no WGD",
  AMP_thresh_tcn = 6)

mergeTN_function(tumor_counts, normal_counts, outfile,
  MINCOV_NORMAL = 25)
}
\arguments{
\item{gene_level}{numeric first argument}

\item{AMP_thresh_tcn}{numeric first argument}

\item{maf}{numeric first argument}

\item{output_pdf}{numeric first argument}

\item{log}{numeric first argument}

\item{threshold}{numeric first argument}

\item{type}{numeric first argument}

\item{purity}{numeric first argument}

\item{nbins}{numeric first argument}

\item{label_WGD}{numeric first argument}

\item{tumor_counts}{numeric first argument}

\item{normal_counts}{numeric first argument}

\item{outfile}{numeric first argument}

\item{MINCOV_NORMAL}{numeric first argument}

\item{AMP_thresh_tcn}{numeric first argument}
}
\value{
don't know

don't know

don't know
}
\description{
don't know

don't know

don't know
}
